ExpressionParser: align parser and expression node filenames
