+++ src/App/Document.cpp
@@ -75,7 +75,7 @@
 #include "Application.h"
 #include "AutoTransaction.h"
 #include "BackupPolicy.h"
-#include "ExpressionParser.h"
+#include "ExpressionNodes.h"
 #include "GeoFeature.h"
 #include "License.h"
 #include "Link.h"
