+++ src/Mod/Spreadsheet/Gui/DlgSheetConf.cpp
@@ -26,7 +26,7 @@
 #include <QMessageBox>
 
 #include <App/Document.h>
-#include <App/ExpressionParser.h>
+#include <App/ExpressionNodes.h>
 #include <App/Range.h>
 #include <Base/Tools.h>
 #include <Gui/CommandT.h>
