+++ src/Mod/Spreadsheet/Gui/PropertiesDialog.cpp
@@ -24,7 +24,7 @@
 
 
 #include <App/Document.h>
-#include <App/ExpressionParser.h>
+#include <App/ExpressionNodes.h>
 #include <App/Range.h>
 #include <Base/Tools.h>
 #include <Gui/CommandT.h>
